pyproject.toml
[project]
name = "gm-favicon-tab-change"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "pillow>=11.2.1",
]
